Fix crashed due to bad role allocation for network bandwidth.
Network bandwidth resource is appended in pre-reservation-refinement
format before being validated and normalized in
post-reservation-refinement format just after the network bandwidth
enforcement.

The crashes were due to the use of Resources class while
using pre-reservation-refinement resources which is not supposed to
happen.

Also, the allocation_info attribute of resources is meant to be set by
Mesos and has been introduced to support multi-role capability for
schedulers.